+++ buildchain/tailwind.config.ts
@@ -1,0 +1,9 @@
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: [
+    '../src/templates/**/*.{twig,html}',
+    './src/vue/**/*.{vue,html}',
+  ],
+  theme: {},
+  plugins: [],
+};