+++ src/helpers/Sitemap.php
@@ -5,6 +5,7 @@ namespace nystudio107\seomatic\helpers;
 use Craft;
 use craft\base\Element;
 use craft\base\ElementInterface;
+use craft\base\Event;
 use craft\console\Application as ConsoleApplication;
 use craft\db\Paginator;
 use craft\elements\Asset;
@@ -13,6 +14,7 @@ use craft\fields\Assets as AssetsField;
 use craft\models\SiteGroup;
 use nystudio107\fastcgicachebust\FastcgiCacheBust;
 use nystudio107\seomatic\base\SeoElementInterface;
+use nystudio107\seomatic\events\IncludeSitemapEntryEvent;
 use nystudio107\seomatic\fields\SeoSettings;
 use nystudio107\seomatic\helpers\Field as FieldHelper;
 use nystudio107\seomatic\jobs\GenerateSitemap;
@@ -30,6 +32,29 @@ use yii\helpers\Html;
  */
 class Sitemap
 {
+    /**
+     * @event IncludeSitemapEntryEvent The event that is triggered when an entry is
+     * about to be included in a sitemap
+     *
+     * ---
+     * ```php
+     * use nystudio107\seomatic\events\IncludeSitemapEntryEvent;
+     * use nystudio107\seomatic\helpers\Sitemap;
+     * use yii\base\Event;
+     * Event::on(Sitemap::class, Sitemap::EVENT_INCLUDE_SITEMAP_ENTRY, function(IncludeSitemapEntryEvent $e) {
+     *     $e->include = false;
+     * });
+     * ```
+     */
+    const EVENT_INCLUDE_SITEMAP_ENTRY = 'includeSitemapEntry';
+
+    /**
+     * Generate a sitemap with the passed in $params
+     *
+     * @param array $params
+     * @return void
+     * @throws \craft\errors\SiteNotFoundException
+     */
     public static function generateSitemap(array $params)
     {
         /** @var \craft\queue\Queue $queue */
@@ -169,8 +194,16 @@ class Sitemap
                         $robotsEnabled = $metaBundle->metaGlobalVars->robots !== 'none' &&
                             $metaBundle->metaGlobalVars->robots !== 'noindex';
                     }
+                    $enabled = $element->getEnabledForSite($metaBundle->sourceSiteId);
+                    $enabled = $enabled && $path !== null && $metaBundle->metaSitemapVars->sitemapUrls && $robotsEnabled;
+                    $event = new IncludeSitemapEntryEvent([
+                        'include' => $enabled,
+                        'element' => $element,
+                        'metaBundle' => $metaBundle,
+                    ]);
+                    Event::trigger(self::class, self::EVENT_INCLUDE_SITEMAP_ENTRY, $event);
                     // Only add in a sitemap entry if it meets our criteria
-                    if ($path !== null && $metaBundle->metaSitemapVars->sitemapUrls && $robotsEnabled) {
+                    if ($event->include) {
                         // Get the url and canonicalUrl
                         try {
                             $url = UrlHelper::siteUrl($path, null, null, $metaBundle->sourceSiteId);
@@ -233,9 +266,10 @@ class Sitemap
                                             $altSourceSiteId
                                         );
                                         if ($altMetaBundle) {
+                                            $altEnabled = $altElement->getEnabledForSite($altMetaBundle->sourceSiteId);
                                             // Make sure this entry isn't disabled
                                             self::combineFieldSettings($altElement, $altMetaBundle);
-                                            if ($altMetaBundle->metaSitemapVars->sitemapUrls) {
+                                            if ($altEnabled && $altMetaBundle->metaSitemapVars->sitemapUrls) {
                                                 try {
                                                     $altUrl = UrlHelper::siteUrl($altElement->url, null, null, $altSourceId);
                                                 } catch (Exception $e) {