+++ src/models/Settings.php
@@ -198,12 +198,17 @@ class Settings extends VarsModel
     public bool $generatorEnabled = true;
 
     /**
-     * @var string
+     * @var string|array
      * SEOmatic uses the Craft `siteUrl` to generate the external URLs.  If you
      * are using it in a non-standard environment, such as a headless GraphQL or
      * ElementAPI server, you can override what it uses for the `siteUrl` below.
-     */
-    public string $siteUrlOverride = '';
+     * This can be either a simple string, or an array of strings indexed by the site
+     * handle, for multi-site setups. e.g.:
+     * 'siteUrlOverride' => [
+     *     'default' => 'http://example.com/',
+     *     'spanish' => 'http://example.com/es/',
+     * ],     */
+    public string|array $siteUrlOverride = '';
 
     /**
      * @var int
@@ -315,7 +320,7 @@ class Settings extends VarsModel
             ['maxTitleLength', 'default', 'value' => 70],
             ['maxDescriptionLength', 'integer', 'min' => 10],
             ['maxDescriptionLength', 'default', 'value' => 155],
-            ['siteUrlOverride', 'string'],
+            ['siteUrlOverride', 'safe'],
             ['siteUrlOverride', 'default', 'value' => ''],
             [
                 [