+++ src/services/MetaContainers.php
@@ -20,6 +20,7 @@ use craft\elements\GlobalSet;
 use nystudio107\seomatic\base\MetaContainer;
 use nystudio107\seomatic\base\MetaItem;
 use nystudio107\seomatic\events\InvalidateContainerCachesEvent;
+use nystudio107\seomatic\events\MetaBundleDebugDataEvent;
 use nystudio107\seomatic\helpers\ArrayHelper;
 use nystudio107\seomatic\helpers\DynamicMeta as DynamicMetaHelper;
 use nystudio107\seomatic\helpers\Field as FieldHelper;
@@ -96,6 +97,22 @@ class MetaContainers extends Component
      */
     const EVENT_INVALIDATE_CONTAINER_CACHES = 'invalidateContainerCaches';
 
+    /**
+     * @event MetaBundleDebugDataEvent The event that is triggered to record MetaBundle
+     * debug data
+     *
+     * ---
+     * ```php
+     * use nystudio107\seomatic\events\MetaBundleDebugDataEvent;
+     * use nystudio107\seomatic\services\MetaContainers;
+     * use yii\base\Event;
+     * Event::on(MetaContainers::class, MetaContainers::EVENT_METABUNDLE_DEBUG_DATA, function(MetaBundleDebugDataEvent $e) {
+     *     // Do something with the MetaBundle debug data
+     * });
+     * ```
+     */
+    const EVENT_METABUNDLE_DEBUG_DATA = 'metaBundleDebugData';
+
     // Public Properties
     // =========================================================================
 
@@ -124,13 +141,13 @@ class MetaContainers extends Component
      */
     public $cachedJsonLdNonce;
 
-    // Protected Properties
-    // =========================================================================
-
     /**
      * @var MetaContainer
      */
-    protected $metaContainers = [];
+    public $metaContainers = [];
+
+    // Protected Properties
+    // =========================================================================
 
     /**
      * @var null|MetaBundle
@@ -196,6 +213,20 @@ class MetaContainers extends Component
             if (Seomatic::$settings->addHrefLang && Seomatic::$settings->addPaginatedHreflang) {
                 DynamicMetaHelper::addMetaLinkHrefLang();
             }
+        }
+        // Fire an 'metaBundleDebugData' event
+        if ($this->hasEventHandlers(self::EVENT_METABUNDLE_DEBUG_DATA)) {
+            $metaBundle = new MetaBundle([
+                'metaGlobalVars' => clone $this->metaGlobalVars,
+                'metaSiteVars' => clone $this->metaSiteVars,
+                'metaSitemapVars' => clone $this->metaSitemapVars,
+                'metaContainers' => $this->metaContainers,
+            ]);
+            $event = new MetaBundleDebugDataEvent([
+                'metaBundleCategory' => MetaBundleDebugDataEvent::COMBINED_META_BUNDLE,
+                'metaBundle' => $metaBundle,
+            ]);
+            $this->trigger(self::EVENT_METABUNDLE_DEBUG_DATA, $event);
         }
         // Add in our http headers
         DynamicMetaHelper::includeHttpHeaders();
@@ -380,13 +411,14 @@ class MetaContainers extends Component
                 ],
             ]);
             $this->containerDependency = $dependency;
-            if (Seomatic::$previewingMetaContainers) {
+            $debugModule = Craft::$app->getModule('debug');
+            if (Seomatic::$previewingMetaContainers || $debugModule) {
                 Seomatic::$plugin->frontendTemplates->loadFrontendTemplateContainers($siteId);
                 $this->loadGlobalMetaContainers($siteId);
                 $this->loadContentMetaContainers();
                 $this->loadFieldMetaContainers();
                 // We only need the dynamic data for headless requests
-                if (Seomatic::$headlessRequest || Seomatic::$plugin->helper::isPreview()) {
+                if (Seomatic::$headlessRequest || Seomatic::$plugin->helper::isPreview() || $debugModule) {
                     DynamicMetaHelper::addDynamicMetaToContainers($uri, $siteId);
                 }
             } else {
@@ -457,6 +489,14 @@ class MetaContainers extends Component
         }
         $metaBundle = Seomatic::$plugin->metaBundles->getGlobalMetaBundle($siteId);
         if ($metaBundle) {
+            // Fire an 'metaBundleDebugData' event
+            if ($this->hasEventHandlers(self::EVENT_METABUNDLE_DEBUG_DATA)) {
+                $event = new MetaBundleDebugDataEvent([
+                    'metaBundleCategory' => MetaBundleDebugDataEvent::GLOBAL_META_BUNDLE,
+                    'metaBundle' => $metaBundle,
+                ]);
+                $this->trigger(self::EVENT_METABUNDLE_DEBUG_DATA, $event);
+            }
             // Meta global vars
             $this->metaGlobalVars = clone $metaBundle->metaGlobalVars;
             // Meta site vars
@@ -853,10 +893,26 @@ class MetaContainers extends Component
                 ?? Craft::$app->getSites()->primarySite->id
                 ?? 1;
         }
+        $element = null;
         $uri = trim($uri, '/');
         /** @var Element $element */
         $enabledOnly = !Seomatic::$previewingMetaContainers;
-        $element = Craft::$app->getElements()->getElementByUri($uri, $siteId, $enabledOnly);
+        // Try to use Craft's matched element if looking for an enabled element, the current `siteId` is being used and
+        // the current `uri` matches what was in the request
+        $request = Craft::$app->getRequest();
+        if ($enabledOnly && !$request->getIsConsoleRequest()) {
+            try {
+                if ($siteId === Craft::$app->getSites()->currentSite->id
+                    && $request->getPathInfo() === $uri) {
+                    $element = Craft::$app->getUrlManager()->getMatchedElement();
+                }
+            } catch (\Throwable $e) {
+                Craft::error($e->getMessage(), __METHOD__);
+            }
+        }
+        if (!$element) {
+            $element = Craft::$app->getElements()->getElementByUri($uri, $siteId, $enabledOnly);
+        }
         if ($element && ($element->uri !== null)) {
             Seomatic::setMatchedElement($element);
         }
@@ -899,6 +955,14 @@ class MetaContainers extends Component
         Craft::beginProfile('MetaContainers::loadContentMetaContainers', __METHOD__);
         $metaBundle = $this->getMatchedMetaBundle();
         if ($metaBundle) {
+            // Fire an 'metaBundleDebugData' event
+            if ($this->hasEventHandlers(self::EVENT_METABUNDLE_DEBUG_DATA)) {
+                $event = new MetaBundleDebugDataEvent([
+                    'metaBundleCategory' => MetaBundleDebugDataEvent::CONTENT_META_BUNDLE,
+                    'metaBundle' => $metaBundle,
+                ]);
+                $this->trigger(self::EVENT_METABUNDLE_DEBUG_DATA, $event);
+            }
             $this->addMetaBundleToContainers($metaBundle);
         }
         Craft::endProfile('MetaContainers::loadContentMetaContainers', __METHOD__);
@@ -955,6 +1019,14 @@ class MetaContainers extends Component
                             $schemaType = MetaValueHelper::parseString($schemaType);
                             $generalContainer->data['mainEntityOfPage'] = MetaJsonLd::create($schemaType, $config);
                         }
+                    }
+                    // Fire an 'metaBundleDebugData' event
+                    if ($this->hasEventHandlers(self::EVENT_METABUNDLE_DEBUG_DATA)) {
+                        $event = new MetaBundleDebugDataEvent([
+                            'metaBundleCategory' => MetaBundleDebugDataEvent::FIELD_META_BUNDLE,
+                            'metaBundle' => $metaBundle,
+                        ]);
+                        $this->trigger(self::EVENT_METABUNDLE_DEBUG_DATA, $event);
                     }
                     $this->addMetaBundleToContainers($metaBundle);
                 }